+++ compiler/lib/reserved.ml
@@ -140,4 +140,5 @@ let provided =
     ; "parseInt"
     ; "module"
     ; "require"
+    ; "Symbol"
     ]